o/devicestate: move system seeds related helpers to separate file
